backend/pipelines/scene-bridge-mvp/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
numpy
trimesh
scipy
viser
